+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for Preview and Preprod Cardano networks via `--testnet-magic` option. Project ID for these networks is picked from environment variables `BLOCKFROST_PROJECT_ID_PREPROD` and `BLOCKFROST_PROJECT_ID_PREVIEW`.
 - Report `cost_models` in protocol-parameters command
+- Fallback to cardano-cli for commands not supported by blockfrost-cardano-cli.
 
 ### Changed